server/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
distro==1.4.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.5.1
msgpack==0.6.2
packaging==20.3
parsel==1.6.0
pep517==0.8.2
Pillow==7.1.2
progress==1.5
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
queuelib==1.5.0
requests==2.22.0
retrying==1.3.3
Scrapy==2.1.0
scrapy-djangoitem==1.1.1
scrapy-user-agents==0.1.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
ua-parser==0.10.0
urllib3==1.25.8
user-agents==2.1
w3lib==1.22.0
webencodings==0.5.1
zope.interface==5.1.0
